Replace status hide flag with a more generic fixed status flag. It takes a string and displays that value.
Also added a statusHue flag to control the hue. They can be used independently, through a fixed status value will always have a fixed hue, falling back to a default of 60 if no other value is set.

To replicate previous hideStatus flag the following AE-like can be used:
{"key":"ActiveEffectLike","mode":"override","path":"flags.pf2e.statusEntry","value":"?"}

The new hue is used as follows:
{"key":"ActiveEffectLike","mode":"override","path":"flags.pf2e.statusHue","value":130}